app/requirements.txt
openai==1.13.3
ruamel.yaml